apis/pyproject.toml
[tool.poetry]
name = "apis"
version = "0.1.0"
description = ""
authors = ["awsconsultant9"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.0"
uvicorn = "^0.35.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.41"
alembic = "^1.16.4"
redis = "^6.2.0"
limits = "^5.4.0"
python-jose = "^3.5.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
